refactor: Refactor: split orbit-store sqlite/task_registry.rs into focu… [ORB-00118]
Refactor: split orbit-store sqlite/task_registry.rs into focused submodules

Planned-By: claude
